Created factory for modal
