+++ index.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html>
+<html lang="fr">
   <head>
     <link href="css/main.css" rel="stylesheet" type="text/css" />
     <meta charset="utf-8">
