+++ photographer.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html>
+<html lang="fr">
 
 <head>
   <link href="css/main.css" rel="stylesheet" type="text/css" />
@@ -16,21 +16,6 @@
     <section class="photograph-header">
     </section>
   </main>
-  <div id="contact_modal">
-    <div class="modal">
-      <header>
-        <h2>Contactez-moi</h2>
-        <img class="close_modal_button" src="assets/icons/close.svg" />
-      </header>
-      <form>
-        <div>
-          <label>Prénom</label>
-          <input />
-        </div>
-        <button class="contact_button">Envoyer</button>
-      </form>
-    </div>
-  </div>
   <script src="/scripts/pages/photographer.js" type="module"></script>
 </body>
